# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: PZIO Task Management System
  description: Backend REST API for the team-project task management system. See documents/4.pdf (SAD)
    for the canonical API contract.
  version: 0.1.0
paths:
...
  /api/auth/register:
    post:
      tags:
      - Auth
      summary: Register a new user
      description: Creates a user with role `TeamMember`. Returns the user without the password hash.
      operationId: register_api_auth_register_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreate'
        required: true
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserRead'
        '400':
          description: Validation error
        '409':
          description: Email already in use
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/auth/login:
    post:
      tags:
      - Auth
      summary: Log in with email and password
      description: 'Returns a JWT access token. Use it as `Authorization: Bearer <token>` on protected
        routes.'
      operationId: login_api_auth_login_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
        required: true
      responses:
        '200':
          description: Authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401':
          description: Invalid credentials
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/users/me:
    get:
      tags:
      - Auth
      summary: Get current user profile
      description: Returns the profile of the currently authenticated user.
      operationId: get_profile_api_users_me_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserRead'
      security:
      - HTTPBearer: []
    patch:
      tags:
      - Auth
      summary: Update current user profile
      description: Updates the profile of the currently authenticated user. All fields are optional.
      operationId: update_profile_api_users_me_patch
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdate'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserRead'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
  /api/users:
    get:
      tags:
      - Auth
      summary: List users (Admin)
      description: Returns a paginated list of users. Requires Administrator role.
      operationId: list_users_api_users_get
      security:
      - HTTPBearer: []
      parameters:
      - name: search
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Search by email, first name, or last name
          title: Search
        description: Search by email, first name, or last name
      - name: isActive
        in: query
        required: false
        schema:
          anyOf:
          - type: boolean
          - type: 'null'
          description: Filter by active status
          title: Isactive
        description: Filter by active status
      - name: sortBy
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Field to sort by
          title: Sortby
        description: Field to sort by
      - name: sortDirection
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Sort direction (asc or desc)
          default: desc
          title: Sortdirection
        description: Sort direction (asc or desc)
      - name: page
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          description: Page number
          default: 1
          title: Page
        description: Page number
      - name: size
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          description: Items per page
          default: 50
          title: Size
        description: Items per page
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedUserResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/users/{id}/status:
    patch:
      tags:
      - Auth
      summary: Change user status (Admin)
      description: Activates or deactivates a user account. Requires Administrator role.
      operationId: update_user_status_api_users__id__status_patch
      security:
      - HTTPBearer: []
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          title: Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserStatusUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserRead'
        '404':
          description: User not found
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/users/{id}/role:
    patch:
      tags:
      - Auth
      summary: Change user role (Admin)
      description: Promotes or demotes a user to another role. Requires Administrator role. Admin cannot
        change their own role.
      operationId: update_user_role_api_users__id__role_patch
      security:
      - HTTPBearer: []
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          title: Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRoleUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserRead'
        '400':
          description: Cannot change own role
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - Administrator role required
        '404':
          description: User not found
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/auth/reset-password:
    post:
      tags:
      - Auth
      summary: Request a password reset
      description: Sends a password reset link to the provided email if the account exists.
      operationId: request_password_reset_api_auth_reset_password_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordResetRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/auth/reset-password/confirm:
    post:
      tags:
      - Auth
      summary: Confirm password reset
      description: Sets a new password using a valid reset token.
      operationId: confirm_password_reset_api_auth_reset_password_confirm_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordResetConfirm'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '400':
          description: Invalid or expired token
        '404':
          description: User not found
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/auth/oauth:
    post:
      tags:
      - Auth
      summary: Log in with OAuth provider
      description: Authenticates a user via external provider (Google/GitHub) and returns a JWT.
      operationId: oauth_login_api_auth_oauth_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OAuthLoginRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '400':
          description: Unsupported provider
        '401':
          description: Invalid OAuth token
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    LoginRequest:
      properties:
        email:
          type: string
          format: email
          title: Email
        password:
          type: string
          maxLength: 128
          minLength: 1
          title: Password
      type: object
      required:
      - email
      - password
      title: LoginRequest
      description: Body for `POST /api/auth/login` (SAD §4.1).
      example:
        email: alice@example.com
        password: correct-horse-battery-staple
    MessageResponse:
      properties:
        message:
          type: string
          title: Message
      type: object
      required:
      - message
      title: MessageResponse
      description: Generic message response.
      example:
        message: Password reset email sent.
    OAuthLoginRequest:
      properties:
        provider:
          type: string
          title: Provider
          description: e.g., 'google' or 'github'
        oauthToken:
          type: string
          title: Oauthtoken
      type: object
      required:
      - provider
      - oauthToken
      title: OAuthLoginRequest
      description: Body for POST /api/auth/oauth
      example:
        oauthToken: ya29.A0AfH6SMB...redacted...
        provider: google
...
    PaginatedUserResponse:
      properties:
        items:
          items:
            $ref: '#/components/schemas/UserRead'
          type: array
          title: Items
        total:
          type: integer
          title: Total
        page:
          type: integer
          title: Page
        size:
          type: integer
          title: Size
      type: object
      required:
      - items
      - total
      - page
      - size
      title: PaginatedUserResponse
      description: Response for `GET /api/users` (Admin only).
      example:
        items:
        - createdAt: '2026-05-14T09:30:00Z'
          email: alice@example.com
          firstName: Alice
          isActive: true
          lastName: Smith
          role: TeamMember
          userId: 42
        page: 1
        size: 20
        total: 1
    PasswordResetConfirm:
      properties:
        token:
          type: string
          title: Token
        newPassword:
          type: string
          maxLength: 128
          minLength: 8
          title: Newpassword
      type: object
      required:
      - token
      - newPassword
      title: PasswordResetConfirm
      description: Body for POST /api/auth/reset-password/confirm
      example:
        newPassword: new-correct-horse-battery-staple
        token: 8c1b0f7e-3d2a-4b9a-9e6f-1a2b3c4d5e6f
    PasswordResetRequest:
      properties:
        email:
          type: string
          format: email
          title: Email
      type: object
      required:
      - email
      title: PasswordResetRequest
      description: Body for POST /api/auth/reset-password
      example:
        email: alice@example.com
...
    TokenResponse:
      properties:
        accessToken:
          type: string
          title: Accesstoken
        tokenType:
          type: string
          title: Tokentype
          default: bearer
        expiresIn:
          type: integer
          title: Expiresin
      type: object
      required:
      - accessToken
      - expiresIn
      title: TokenResponse
      description: Response for login / oauth (SAD §4.1).
      example:
        accessToken: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiI0MiIsInJvbGUiOiJUZWFtTWVtYmVyIiwiZXhwIjoxNzc4Nzk1NDIwfQ.signature
        expiresIn: 3600
        tokenType: bearer
    UserCreate:
      properties:
        email:
          type: string
          format: email
          title: Email
        password:
          type: string
          maxLength: 128
          minLength: 8
          title: Password
        firstName:
          type: string
          maxLength: 100
          minLength: 1
          title: Firstname
        lastName:
          type: string
          maxLength: 100
          minLength: 1
          title: Lastname
      type: object
      required:
      - email
      - password
      - firstName
      - lastName
      title: UserCreate
      description: Body for `POST /api/auth/register` (SAD §4.1).
      example:
        email: alice@example.com
        firstName: Alice
        lastName: Smith
        password: correct-horse-battery-staple
    UserRead:
      properties:
        userId:
          type: integer
          title: Userid
        email:
          type: string
          format: email
          title: Email
        firstName:
          type: string
          title: Firstname
        lastName:
          type: string
          title: Lastname
        avatar:
          anyOf:
          - type: string
          - type: 'null'
          title: Avatar
        role:
          $ref: '#/components/schemas/UserRole'
        isActive:
          type: boolean
          title: Isactive
        createdAt:
          type: string
          format: date-time
          title: Createdat
      type: object
      required:
      - userId
      - email
      - firstName
      - lastName
      - role
      - isActive
      - createdAt
      title: UserRead
      description: Public user representation. Never includes `passwordHash` (NFR04).
      example:
        createdAt: '2026-05-14T09:30:00Z'
        email: alice@example.com
        firstName: Alice
        isActive: true
        lastName: Smith
        role: TeamMember
        userId: 42
    UserRole:
      type: string
      enum:
      - TeamMember
      - Manager
      - Administrator
      title: UserRole
      description: 'Roles per System Analysis Document §4 (UML class diagram).


        "Guest" from the SRS actor table is intentionally omitted: it describes the

        unauthenticated visitor and has no corresponding row in the users table.'
    UserRoleUpdate:
      properties:
        role:
          $ref: '#/components/schemas/UserRole'
      type: object
      required:
      - role
      title: UserRoleUpdate
      description: Body for `PATCH /api/users/{id}/role` (Admin only).
      example:
        role: Manager
    UserStatusUpdate:
      properties:
        isActive:
          type: boolean
          title: Isactive
      type: object
      required:
      - isActive
      title: UserStatusUpdate
      description: Body for `PATCH /api/users/{id}/status` (Admin only).
      example:
        isActive: false
    UserUpdate:
      properties:
        firstName:
          anyOf:
          - type: string
            maxLength: 100
            minLength: 1
          - type: 'null'
          title: Firstname
        lastName:
          anyOf:
          - type: string
            maxLength: 100
            minLength: 1
          - type: 'null'
          title: Lastname
        avatar:
          anyOf:
          - type: string
            maxLength: 255
          - type: 'null'
          title: Avatar
      type: object
      title: UserUpdate
      description: Body for `PATCH /api/users/me`. All fields are optional.
      example:
        avatar: https://cdn.example.com/avatars/42.png
        firstName: Alice
        lastName: Smith-Jones
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer
...